update: added a duplicate feature

+++ src/services/apiCabins.js
@@ -40,9 +40,12 @@ export const createEditCabin = async (newCabin, id) => {
     throw new Error("Cabins could not be created");
   }
 
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin?.image);
+
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data?.id);
     console.error(storageError);
